Few changes and some additions
Adjusted the height of the service boxes on the home page to stop them
spazzing out and added the code you gave to me to index.php. Also had a
good go at the services page. Its pulling in the content from the pages
perfectly so just need to sort out the images, mobile layout and look at
giving the services page its own excerpt length.