Some Changes and Additions
Added service-page.php which should be the template for the design,
develop and host pages.

Replaced dummy text on front-page.php with real text but you might want
to change that.

Fixed a few links like the Contact Us Now! buttons (hope I did them
right).

Also there are a couple of paragraphs in design content.txt that may or
may not be useful.