Changes relating to the services page
Added Function to allow excerpts on pages.
Added new images and updated others.
Tweaked a few of the colours.
Improved the SCSS for the services page.
Added a counter and a check determine odd and even objects in the loop.
Means that the text direction will alternate as pages are pulled in to
the page.